+++ packages/project-template-browser-react/src/components/Introduction/Introduction.test.tsx
@@ -1,5 +1,5 @@
+import { expectElementSnapshot } from '@easy/utils-test';
 import { render } from '@testing-library/react';
-import renderer from 'react-test-renderer';
 
 import { Introduction } from './Introduction';
 
@@ -7,9 +7,7 @@ describe('Test component `Introduction`', () => {
     it('renders', () => {
         const reactElement = <Introduction />;
 
-        const component = renderer.create(reactElement);
-        const tree = component.toJSON();
-        expect(tree).toMatchSnapshot();
+        expectElementSnapshot(reactElement);
 
         const { getByText } = render(reactElement);
         const linkElement = getByText(/Learn easy-projs/i);