+++ packages/project-template-browser-react/src/redux/actions/reducers.DisplayConfig.test.ts
@@ -1,11 +1,10 @@
+import { expectSnapshot } from '@easy/utils-test';
 import { getInitialState, _actions, _reducer } from './slice';
 
 describe('Test redux reducer `updateClientAreaSize`', () => {
     it('updates client area size', () => {
         const prevState = getInitialState();
 
-        expect(
-            _reducer(prevState, _actions.updateClientAreaSize({ height: 123, width: 456 })).displayConfig,
-        ).toMatchSnapshot();
+        expectSnapshot(_reducer(prevState, _actions.updateClientAreaSize({ height: 123, width: 456 })).displayConfig);
     });
 });