+++ packages/project-template-browser-react/src/redux/actions/reducers.test.ts
@@ -1,9 +1,11 @@
+import { expectSnapshot } from '@easy/utils-test';
+
 import { getInitialState, _actions, _reducer } from './slice';
 
 describe('Test redux reducer `_finishPreparing`', () => {
     it('finishes preparing', () => {
         const prevState = getInitialState();
 
-        expect(_reducer(prevState, _actions._finishPreparing()).prepared).toMatchSnapshot();
+        expectSnapshot(_reducer(prevState, _actions._finishPreparing()).prepared);
     });
 });